........
  r36981 | res2002 | 2011-10-10 11:54:47 +0200 (lun., 10 oct. 2011) | 1 line
  
  Removed unused code from xmltiny plugin
........
  r36982 | res2002 | 2011-10-10 11:54:59 +0200 (lun., 10 oct. 2011) | 1 line
  
  Changed the CS_BUILD_IFELSE autoconf macro to to put the tested and 'other' cflags _after_ the default cflags: this seems to better reflect how additional cflags are commonly used: after the default flags. This is relevant for flags which are order-sensitive, e.g. g++ the -std= option.
........
  r36987 | res2002 | 2011-10-10 11:56:28 +0200 (lun., 10 oct. 2011) | 1 line
  
  Warning fix
........
  r36988 | res2002 | 2011-10-10 11:56:39 +0200 (lun., 10 oct. 2011) | 1 line
  
  Don't use separate sections in 'profile' build mode
........
  r37000 | res2002 | 2011-10-20 18:27:46 +0200 (jeu., 20 oct. 2011) | 1 line
  
  Fix use of _AC_CC autoconf macro
........
  r37001 | res2002 | 2011-10-20 18:28:00 +0200 (jeu., 20 oct. 2011) | 1 line
  
  Only perform separate sections check once
........
  r37020 | jorrit | 2011-10-25 12:12:06 +0200 (mar., 25 oct. 2011) | 3 lines
  
  Fixed a dependency problem between WXGL canvas and font server plugin.
........
  r37024 | jorrit | 2011-10-28 09:24:03 +0200 (ven., 28 oct. 2011) | 5 lines
  
  Fixed a problem with the WXGL canvas where it would send out uppercase
  chars in case a lowercase key was pressed. The solution is a bit of a hack
  though but I couldn't find a better solution.
........
  r37025 | jorrit | 2011-10-28 10:07:39 +0200 (ven., 28 oct. 2011) | 3 lines
  
  Implemented SetMousePosition() for WXGL canvas.
........


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/release/V2.0@37078 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shadercompiler/xmlshader/cpi/condition.h
@@ -173,7 +173,7 @@ class csHashComputer<NS_XMLSHADER::CondOperation>
           (char*)(tempOp.left.svLocation.indices + 1),
           *tempOp.left.svLocation.indices * sizeof (size_t));
       }
-      tempOp.left.svLocation.indices = (size_t*)leftIndexHash;
+      tempOp.left.svLocation.indices = (size_t*)(uintptr_t)leftIndexHash;
     }
     tempOp.right = right;
     if (tempOp.right.type >= NS_XMLSHADER::operandSV)
@@ -184,7 +184,7 @@ class csHashComputer<NS_XMLSHADER::CondOperation>
           (char*)(tempOp.right.svLocation.indices + 1),
           *tempOp.right.svLocation.indices * sizeof (size_t));
       }
-      tempOp.right.svLocation.indices = (size_t*)rightIndexHash;
+      tempOp.right.svLocation.indices = (size_t*)(uintptr_t)rightIndexHash;
     }
     return csHashCompute ((char*)&tempOp, sizeof (tempOp));
   }
